var-default-image-swu: Add default sw-description and symbolic link to it
This fixes:
WARNING: sources/meta-variscite-ti-sdk/dynamic-layers/swupdate/var-default-image-swu.bb: Unable to get checksum for var-default-image-swu SRC_URI entry sw-description: file could not be found

The warning is shown during parsing when searching for files for other
machines:
- am62x-var-som-k3r5-gp
- am62x-var-som-k3r5-hs-fs
- am62x-var-som-k3r5-hs-se

sw-description is not needed for these machines. so rather than creating
directories for each machine, just create a default file and symbolic
link to it for am62x-var-som machine.

Signed-off-by: Nate Drude <nate.d@variscite.com>