perl: skip shebang-size QA check for nativesdk-perl packages
When nativesdk-perl is built as part of meta-toolchain-arago-tisdk,
the following error is raised:
maximum shebang size exceeded, the maximum size is 128.
These changes have been ported from the following commit ID
https://git.ti.com/cgit/arago-project/meta-arago/commit/?id=fa0e7b7f0c944555c54d7698dbfe630f3c454301
in Kirstone branch, where have been later removed with
https://git.ti.com/cgit/arago-project/meta-arago/commit/?id=ed2ba127e016fe0fca93ae863396b79719ab8db3
However, this is a workarond specific for Dunfell and should
not be included in the upcoming Kirkstone branch.

Signed-off-by: Pierluigi Passaro <pierluigi.p@variscite.com>