openssh: break dependency on base package for -dev package
When building the SDK using populate-sdk, the following error is raised
package openssh conflicts with dropbear
This problem is actually am335x-var-som specific and not reproducible
using am62x-var-som machine, which is however not affected by the
changes.
The different behaviour is orginated by the different commit ID
of oe-core referenced in Processor SDKs 08.02 and 08.06 BSPs.
These changes have been pushed in oe-core with the following commit
https://github.com/openembedded/openembedded-core/commit/e863fc060940d11cd6fd58f0f314333ed419cf54

Signed-off-by: Pierluigi Passaro <pierluigi.p@variscite.com>