Revert "layer: bbmask meta-arago westoni-init bbappend"
Prior to Kirkstone, the bbappend provided in meta-arago unconditionally
disabled the weston-init systemd service. This was undesirable for
Variscite's system and thus, we masked the append.

However, in [1], TI restructured the append so that it no longer
disables the sytstemd service. Moreover, the weston-socket.sh profile
script installed by the TI append is required for proper Qt and Weston
app launches and we can now use it by unmasking the bbappend.

This reverts commit d0a27b21ae3e913e814667633fda411868286980.

[1] https://git.ti.com/cgit/arago-project/meta-arago/commit/?id=e23f1ef439f04c66abc35255cb3ce0f7684c6213

Signed-off-by: Ken Sloat <ken.s@variscite.com>